+++ Source_Files/TCPMess/CommunicationsChannel.h
@@ -182,12 +182,12 @@ public:
 
 	// Callers can use these (compared with machine_tick_count()) to gauge activity on the Channel:
 	// each time pump() receives/sends new data, value is set to machine_tick_count() at that time.
-	Uint32		ticksAtLastReceive() const { return mTicksAtLastReceive; }
-	Uint32		ticksAtLastSend() const { return mTicksAtLastSend; }
+	uint64_t	ticksAtLastReceive() const { return mTicksAtLastReceive; }
+	uint64_t	ticksAtLastSend() const { return mTicksAtLastSend; }
 
 	// Or callers can just use these.
-	Uint32		millisecondsSinceLastReceive() const { return machine_tick_count() - mTicksAtLastReceive; }
-	Uint32		millisecondsSinceLastSend() const { return machine_tick_count() - mTicksAtLastSend; }
+	uint64_t	millisecondsSinceLastReceive() const { return machine_tick_count() - mTicksAtLastReceive; }
+	uint64_t	millisecondsSinceLastSend() const { return machine_tick_count() - mTicksAtLastSend; }
 
 private:
 	enum CommunicationResult
@@ -227,7 +227,7 @@ private:
 	UninflatedMessage* mIncomingMessage;
 	size_t		mIncomingMessagePosition;
 
-	Uint32		mTicksAtLastReceive;
+	uint64_t	mTicksAtLastReceive;
 
 	typedef std::list<Message*>	MessageQueue;
 	MessageQueue	mIncomingMessages;
@@ -236,7 +236,7 @@ private:
 	Uint8		mOutgoingHeader[kHeaderPackedSize];
 	size_t		mOutgoingHeaderPosition;
 
-	Uint32		mTicksAtLastSend;
+	uint64_t	mTicksAtLastSend;
 
 	typedef std::list<UninflatedMessage*>	UninflatedMessageQueue;
 	UninflatedMessageQueue	mOutgoingMessages;
